File moving and SCRIPT-O-MATIC creation
I deleted some files that I didn’t need anymore because I incorporated
their functionality into my single script, “script-o-matic.” I also
reconfigured the Apple Script App to run the script-o-matic.py file.

For script-o-matic, this is was the initial rollout of the file. In it
to start, I only added the working parts of the Canadian document.